cup_uniforms - reorganize configs into separate folders
