+++ addons/cup_uniforms/include/Nomex/CfgVehicles.hpp
@@ -1,0 +1,21 @@
+class CUP_B_USMC_Soldier_06;
+
+class tmtm_nomex_grey: CUP_B_USMC_Soldier_06
+{
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+
+    displayName="Nomex Flightsuit (Grey)";
+    uniformClass="tmtm_u_nomex_grey";
+
+    picture="\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USMC\data\ui\icon_u_b_usmc_pilotoverall_ca.paa";
+    hiddenSelections[] = {"Camo", "Camo2"};
+    hiddenSelectionsTextures[] =
+    {
+        "\tmtm\pack\cup_uniforms\Data\tmtm_nomex_grey_co.paa",
+        "\cup\creatures\people\military\cup_creatures_people_military_usmc\data\gloves_co.paa"
+    };
+};
