+++ addons/cup_uniforms/include/VKPO/CfgVehicles.hpp
@@ -1,0 +1,1042 @@
+class CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1;
+/*
+hiddenSelectionsTextures[] = {
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_hands_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_VKPO_Pants_s_CO.paa"
+};
+*/
+class tmtm_vkpo_jacket_winter_Forest_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_Forest_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_ForestMountain_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+
+    displayName="VKPO Winter Jacket 1 (CDF Forest/Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_ForestMountain_1";
+
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_ForestOxblood_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+
+    displayName="VKPO Winter Jacket 1 (CDF Forest/Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_ForestOxblood_1";
+
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_ForestMTP_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+
+    displayName="VKPO Winter Jacket 1 (CDF Forest/MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_ForestMTP_1";
+
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_Mountain_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_Mountain_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MountainForest_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Mountain/Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MountainForest_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MountainOxblood_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Mountain/Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MountainOxblood_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MountainMTP_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Mountain/MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MountainMTP_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_Oxblood_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_Oxblood_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_OxbloodForest_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Oxblood/Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_OxbloodForest_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_OxbloodMountain_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Oxblood/Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_OxbloodMountain_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_OxbloodMTP_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Oxblood/MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_OxbloodMTP_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MTP_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MTP_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MTPForest_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (MTP/CDF Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MTPForest_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MTPMountain_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (MTP/CDF Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MTPMountain_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MTPOxblood_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (MTP/CDF Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MTPOxblood_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_blackForest_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (Black/CDF Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_blackForest_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_blackMountain_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (Black/CDF Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_blackMountain_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_blackOxblood_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (Black/CDF Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_blackOxblood_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_blackMTP_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (Black/MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_blackMTP_1";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa"
+    };
+};
+class CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_2;
+/*
+hiddenSelectionsTextures[] = {
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_hands_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_VKPO_Pants_s_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+};
+*/
+class CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3;
+/*
+hiddenSelectionsTextures[] = {
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_hands_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_VKPO_Pants_s_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+};
+*/
+class tmtm_vkpo_jacket_winter_Forest_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_Forest_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_ForestMountain_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+
+    displayName="VKPO Winter Jacket 2 (CDF Forest/Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_ForestMountain_2";
+
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_ForestOxblood_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+
+    displayName="VKPO Winter Jacket 2 (CDF Forest/Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_ForestOxblood_2";
+
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_ForestMTP_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+
+    displayName="VKPO Winter Jacket 2 (CDF Forest/MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_ForestMTP_2";
+
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_Mountain_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_Mountain_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MountainForest_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Mountain/Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MountainForest_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MountainOxblood_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Mountain/Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MountainOxblood_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MountainMTP_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Mountain/MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MountainMTP_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_Oxblood_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_Oxblood_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_OxbloodForest_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Oxblood/Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_OxbloodForest_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_OxbloodMountain_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Oxblood/Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_OxbloodMountain_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_OxbloodMTP_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Oxblood/MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_OxbloodMTP_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MTP_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MTP_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MTPForest_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (MTP/CDF Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MTPForest_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MTPMountain_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (MTP/CDF Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MTPMountain_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MTPOxblood_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (MTP/CDF Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MTPOxblood_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_blackForest_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (Black/CDF Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_blackForest_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_blackMountain_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (Black/CDF Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_blackMountain_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_blackOxblood_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (Black/CDF Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_blackOxblood_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_blackMTP_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (Black/MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_blackMTP_2";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+
+class CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4;
+/*
+hiddenSelectionsTextures[] = {
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_hands_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_VKPO_Pants_s_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+};
+
+*/
+class tmtm_vkpo_jacket_winter_Forest_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_Forest_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_ForestMountain_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+
+    displayName="VKPO Winter Jacket 3 (CDF Forest/Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_ForestMountain_3";
+
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_ForestOxblood_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+
+    displayName="VKPO Winter Jacket 3 (CDF Forest/Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_ForestOxblood_3";
+
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_ForestMTP_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+
+    displayName="VKPO Winter Jacket 3 (CDF Forest/MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_ForestMTP_3";
+
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_Mountain_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_Mountain_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MountainForest_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Mountain/Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MountainForest_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MountainOxblood_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Mountain/Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MountainOxblood_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MountainMTP_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Mountain/MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MountainMTP_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_Oxblood_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_Oxblood_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_OxbloodForest_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Oxblood/Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_OxbloodForest_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_OxbloodMountain_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Oxblood/Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_OxbloodMountain_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_OxbloodMTP_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Oxblood/MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_OxbloodMTP_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MTP_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MTP_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MTPForest_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (MTP/CDF Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MTPForest_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MTPMountain_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (MTP/CDF Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MTPMountain_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_MTPOxblood_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (MTP/CDF Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_MTPOxblood_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_blackForest_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (Black/CDF Forest)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_blackForest_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_blackMountain_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (Black/CDF Mountain)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_blackMountain_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_blackOxblood_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (Black/CDF Oxblood)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_blackOxblood_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+class tmtm_vkpo_jacket_winter_blackMTP_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (Black/MTP)";
+    uniformClass="tmtm_u_vkpo_jacket_winter_blackMTP_3";
+    picture="";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+};
+
+
+class CUP_U_O_RUS_Soldier_VKPO_Winter_1;
+/*
+hiddenSelectionsTextures[] = {
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_hands_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_VKPO_Pants_w_CO.paa"
+};
+*/
+class CUP_U_O_RUS_Soldier_VKPO_Winter_2;
+/*
+hiddenSelectionsTextures[] = { 
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_hands_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_VKPO_Pants_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+};
+*/
+class CUP_U_O_RUS_Soldier_VKPO_Winter_3;
+/*
+hiddenSelectionsTextures[] = {
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_hands_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_VKPO_Pants_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+};
+*/
+class CUP_U_O_RUS_Soldier_VKPO_Winter_4;
+/*
+hiddenSelectionsTextures[] = {
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\VKPO_Winter\data\ra_VKPO_Jacket_hands_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_VKPO_Pants_w_CO.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+    "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+};
+*/
