+++ addons/cup_uniforms/include/VKPO/CfgWeapons.hpp
@@ -1,0 +1,1230 @@
+class CUP_U_O_RUS_Soldier_VKPO_Winter_1; // displayName = "AFRF VKPO (Winter)";
+class CUP_U_O_RUS_Soldier_VKPO_Winter_2; // displayName = "AFRF VKPO (Winter/Gloves/6B51)";
+class CUP_U_O_RUS_Soldier_VKPO_Winter_3; // displayName = "AFRF VKPO (Winter/Gloves/SSO Kh)";
+class CUP_U_O_RUS_Soldier_VKPO_Winter_4; // displayName = "AFRF VKPO (Winter/Gloves/SSO)";
+
+
+class CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1; // displayName = "AFRF VKPO (Winter jacket)";
+class tmtm_u_vkpo_jacket_winter_Forest_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_Forest_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_ForestMountain_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Forest/Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_ForestMountain_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_ForestOxblood_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Forest/Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_ForestOxblood_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_ForestMTP_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Forest/MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_ForestMTP_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_Mountain_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_Mountain_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MountainForest_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Mountain/Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MountainForest_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MountainOxblood_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Mountain/Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MountainOxblood_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MountainMTP_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Mountain/MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MountainMTP_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_Oxblood_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_Oxblood_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_OxbloodForest_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Oxblood/Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_OxbloodForest_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_OxbloodMountain_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Oxblood/Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_OxbloodMountain_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_OxbloodMTP_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (CDF Oxblood/MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_OxbloodMTP_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MTP_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MTP_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MTPForest_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (MTP/CDF Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MTPForest_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MTPMountain_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (MTP/CDF Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MTPMountain_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MTPOxblood_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (MTP/CDF Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MTPOxblood_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_blackForest_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (Black/CDF Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_blackForest_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_blackMountain_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (Black/CDF Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_blackMountain_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_blackOxblood_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (Black/CDF Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_blackOxblood_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_blackMTP_1: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_1 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 1 (Black/MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_blackMTP_1";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_2; // displayName = "AFRF VKPO (Winter jacket/Gloves/6B51)";
+class CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3; // displayName = "AFRF VKPO (Winter jacket/Gloves/SSO Kh)";
+class tmtm_u_vkpo_jacket_winter_Forest_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_Forest_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_ForestMountain_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Forest/Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_ForestMountain_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_ForestOxblood_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Forest/Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_ForestOxblood_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_ForestMTP_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Forest/MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_ForestMTP_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_Mountain_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_Mountain_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MountainForest_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Mountain/Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MountainForest_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MountainOxblood_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Mountain/Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MountainOxblood_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MountainMTP_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Mountain/MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MountainMTP_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_Oxblood_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_Oxblood_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_OxbloodForest_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Oxblood/Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_OxbloodForest_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_OxbloodMountain_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Oxblood/Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_OxbloodMountain_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_OxbloodMTP_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (CDF Oxblood/MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_OxbloodMTP_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MTP_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MTP_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MTPForest_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (MTP/CDF Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MTPForest_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MTPMountain_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (MTP/CDF Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MTPMountain_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MTPOxblood_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (MTP/CDF Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MTPOxblood_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_blackForest_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (Black/CDF Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_blackForest_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_blackMountain_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (Black/CDF Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_blackMountain_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_blackOxblood_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (Black/CDF Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_blackOxblood_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_blackMTP_2: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_3 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 2 (Black/MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_blackMTP_2";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4; // displayName = "AFRF VKPO (Winter jacket/Gloves/SSO)";
+class tmtm_u_vkpo_jacket_winter_Forest_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_Forest_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_ForestMountain_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Forest/Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_ForestMountain_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_ForestOxblood_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Forest/Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_ForestOxblood_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_ForestMTP_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Forest/MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_forest_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_ForestMTP_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_Mountain_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_Mountain_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MountainForest_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Mountain/Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MountainForest_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MountainOxblood_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Mountain/Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MountainOxblood_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MountainMTP_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Mountain/MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mountain_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MountainMTP_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_Oxblood_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_Oxblood_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_OxbloodForest_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Oxblood/Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_OxbloodForest_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_OxbloodMountain_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Oxblood/Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_OxbloodMountain_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_OxbloodMTP_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (CDF Oxblood/MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_oxblood_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_OxbloodMTP_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MTP_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MTP_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MTPForest_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (MTP/CDF Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MTPForest_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MTPMountain_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (MTP/CDF Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MTPMountain_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_MTPOxblood_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (MTP/CDF Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_mtp_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_MTPOxblood_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_blackForest_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (Black/CDF Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_forest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_blackForest_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_blackMountain_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (Black/CDF Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_blackMountain_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_blackOxblood_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (Black/CDF Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_oxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_blackOxblood_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
+class tmtm_u_vkpo_jacket_winter_blackMTP_3: CUP_U_O_RUS_Soldier_VKPO_Jacket_Winter_4 {
+    author="trenchgun";
+    scope=1;
+    scopeArsenal=2;
+    scopeCurator=0;
+    displayName="VKPO Winter Jacket 3 (Black/MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacket_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpoJacketHands_black_co.paa",
+        "\tmtm\pack\cup_uniforms\Data\tmtm_vkpopants_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\ra_Gloves_Shoes_s_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_RussiaModern\Outfit\data\6b51_s_CO.paa"
+    };
+    class ItemInfo: UniformItem
+    {
+        uniformModel="-";
+        uniformClass="tmtm_vkpo_jacket_winter_blackMTP_3";
+        containerClass="Supply40";
+        mass=20;
+    };
+};
